pmm: allocate a large enough bitmap
The bitmap must be large enough to hold the highest address in the
memory map, not the sum of the sizes of the regions. This is because
regions in the memory map might not be contiguous.

Moreover, the total number of frames might not be a multiple of 8. In
that case, we need an extra byte in the bitmap to hold the tail
frames, so use usize::div_ceil() to add that extra byte when needed.