pmm: actually return a physical address from allocate()
Return a physical address instead of a frame number by multiplying by
the frame size. A function to do this already exists, so use it.

While we are at it, use u8::trailing_ones() to find the lowest zero
bit in a byte, which the compiler can better optimize as a single
instruction.