pmm: fix incorrect setting of allocated frames in bitmap
The iterator in PhysicalFrameAllocator::allocate_contiguous() was
wrong, as we need to multiply the number of frames by the frame size
to get the correct end address of the range. This mirrors a similar
loop in PhysicalFrameAllocator::deallocate_contiguous().

Moreover, base is already a physical address, so we do not need to
multiply it by the frame size.